Split functional tests
It will make functional tests easier to maintain
